Move primary module AbcNesstModule to modules folder
