+++ src/main.ts
@@ -1,5 +1,5 @@
 import { NestFactory } from '@nestjs/core';
-import { AbcNestModule } from './abc-nest.module';
+import { AbcNestModule } from './modules/abc-nest.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AbcNestModule);
